Added option to force proxy usage.
Allows users to pass a flag to force the addition
or absence of a proxy prefix. Particularly for ability
to check for 404 in urls without sending everything
through the proxy since 404s are 404s and proxying
may return false negatives of sites that are seen
as just having a wrong proxy configuration rather
than being a broken link.